moving some functions around

+++ deprecated-but-keep/fleet-log-remove.js
@@ -1,10 +1,10 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { deleteShipLog, getAllShipLogs, getAssistantShipLogs, getShipLogByEntryId } = require('../../api/shipLogApi');
-const { getUsers } = require('../../api/userlistApi');
+const { deleteShipLog, getCrewShipLogs, getShipLogByEntryId, getShipLogsByCommanderId } = require('../api/shipLogApi');
+const { getUserById, getUsers } = require('../api/userlistApi');
 
 
 const command = new SlashCommandBuilder()
-    .setName('xmoderator-fleet-log-remove')
+    .setName('fleet-log-remove')
     .setDescription('Remove a kill log for your ship to the Black Box.')
     .addStringOption(option => 
         option.setName('log')
@@ -15,16 +15,6 @@ const command = new SlashCommandBuilder()
 module.exports = {
     data: command,
     async execute(interaction, client, openai) {
-        const member = interaction.member;
-        const moderatorRoles = process.env.LIVE_ENVIRONMENT === "true" ? process.env.MODERATOR_ROLES.split(',') : process.env.TEST_MODERATOR_ROLES.split(',');
-        const hasPermission = member.roles.cache.some(role => moderatorRoles.includes(role.id));
-        
-        if (!hasPermission) {
-            return interaction.reply({ 
-                content: `${interaction.user.username}, you do not have permission to use this command.`, 
-                ephemeral: false 
-            });
-        }
         // Get the needed variables
         const killLog = interaction.options.getString('log');
         // const allBlackBoxLogs = await getBlackBoxesByUserId(interaction.user.id); // Fetch all black box logs
@@ -34,12 +24,19 @@ module.exports = {
             const channelId = process.env.LIVE_ENVIRONMENT === "true" ? process.env.AUDIT_CHANNEL : process.env.TEST_AUDIT_CHANNEL; // Replace with actual channel ID
             const channel = await client.channels.fetch(channelId);
             const logRecord = await getShipLogByEntryId(killLog); // Fetch the kill log record
+            if(interaction.user.id !== logRecord.commander){
+                const originalCreator = await getUserById(logRecord.commander);
+                return interaction.reply({ 
+                    content: `Only ${originalCreator.username} or a Marauder+ can delete this log: (${logRecord.id}).`, 
+                    ephemeral: false 
+                });
+            }
             if (channel && channel.isTextBased()) {
             await channel.send(`The following Ship Log was deleted by ${interaction.user.username}: \n` + JSON.stringify(logRecord));
             }
 
             await deleteShipLog(killLog); // Pass the selected kill log ID
-            await interaction.reply({ content: `Ship Log deleted successfullyby a moderator: ${interaction.user.username}!`, ephemeral: false });
+            await interaction.reply({ content: 'Ship Log deleted successfully!', ephemeral: true });
         } catch (error) {
             console.error(error);
             await interaction.reply({ content: 'There was an error deleting the Ship Log.', ephemeral: true });
@@ -47,20 +44,19 @@ module.exports = {
     },
     async autocomplete(interaction) {
         const focusedValue = interaction.options.getFocused(); // Get the focused option value
-        const allPrimaryBlackBoxLogs = await getAllShipLogs(interaction.user.id); // Fetch all black box logs
-        const allUsers = await getUsers(); // Fetch all users
+        const allFleetLogsCombined = await getShipLogsByCommanderId(interaction.user.id); // Fetch all black box logs
+        const allUsers = await getUsers()
 
         // Combine ship_used and victims[] into a single searchable array
-        const allBlackBoxLogsListed = allPrimaryBlackBoxLogs.map(log => {
-            // Find the commander in the allUsers array
-            const commander = allUsers.find(user => user.id === log.commander);
-            const commanderName = commander ? commander.username : 'Unknown Commander';
-
-            return {
-                name: `(${log.id}) - Commander: ${commanderName}`,
-                value: log.id // Use the log ID as the value for selection
-            };
-        }).sort((a, b) => Number(b.value) - Number(a.value)); // Sort by log.id in descending order
+        const allBlackBoxLogsListed = allFleetLogsCombined.map(log => ({
+            name: `${log.id} - Subcommanders: ${log.subcommanders
+                .map(subId => {
+                    const user = allUsers.find(user => user.id === subId); // Find the user in allUsers
+                    return user ? user.username : 'Unknown'; // Replace with username or 'Unknown' if not found
+                })
+                .join(', ')}`,
+            value: log.id // Use the log ID as the value for selection
+        }));
 
         // Filter logs based on the focused value (search both ship_used and victims)
         const filtered = allBlackBoxLogsListed.filter(log =>
@@ -73,3 +69,8 @@ module.exports = {
         );
     }
 };
+
+async function getPlayerName(allUsers, playerId) {
+    const userData = await getUserById(playerId)
+    return userData ? userData.nickname : userData.username; // Fallback to playerId if not found
+}
